Actualiza

+++ ejercicio7_cloud_microservicio_vuelo/src/main/java/com/vuelo/service/VueloServiceImpl.java
@@ -3,7 +3,11 @@ package com.vuelo.service;
 import com.vuelo.dao.VueloDao;
 import com.vuelo.model.Vuelo;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.RestTemplate;
+import org.springframework.web.server.ResponseStatusException;
+
 import java.util.List;
 import java.util.Optional;
 
@@ -12,6 +16,11 @@ public class VueloServiceImpl implements VueloService {
 
     @Autowired
     private VueloDao vueloDao;
+    
+    @Autowired
+    private RestTemplate restTemplate;
+
+    private static final String RESERVA_SERVICE_URL = "http://reserva-service/reservas/vuelo";
 
     @Override
     public List<Vuelo> findAllAvailable(Integer plazas) {
@@ -58,9 +67,16 @@ public class VueloServiceImpl implements VueloService {
     @Override
     public boolean deleteVuelo(Long id) {
         if (vueloDao.existsById(id)) {
+            // Verificar si hay reservas activas para el vuelo
+            Boolean reservasActivas = restTemplate.getForObject(RESERVA_SERVICE_URL + "/" + id, Boolean.class);
+            if (reservasActivas != null && reservasActivas) {
+                throw new ResponseStatusException(HttpStatus.CONFLICT, "No se puede eliminar el vuelo porque tiene reservas activas.");
+            }
+            // si se puede eliminar el vuelo al no tener reservas activas
             vueloDao.deleteById(id);
             return true;
         }
+        // no se puede eliminar el vuelo al tener reservas activas
         return false;
     }
